# docs/swagger.yaml
definitions:
  common_middleware.Response:
    properties:
      data:
        type: object
      errmsg:
        type: string
      errno:
        type: integer
      stack:
        type: object
      trace_id:
        type: object
    type: object
  dto.ChangeUserPwdReq:
    properties:
      password:
        example: "123456"
        type: string
    required:
    - password
    type: object
...
  dto.CreateOrUpdateGrpcServiceReq:
    properties:
      black_list:
        description: 黑名单ip列表
        type: string
      client_ip_flow_limit:
        description: 客户端ip限流
        type: integer
      forbid_list:
        description: 禁用ip列表
        type: string
      header_transform:
        description: header转换
        type: string
      ip_list:
        description: ip列表
        type: string
      open_auth:
        description: 是否开启权限
        type: integer
      port:
        description: 端口
        type: integer
      round_type:
        description: 轮询方式
        type: integer
      service_desc:
        description: 服务描述
        type: string
      service_host_flow_limit:
        description: 服务端主机限流
        type: integer
      service_name:
        description: 服务名称
        type: string
      weight_list:
        description: 权重列表
        type: string
      white_host_name:
        description: 白名单主机列表
        type: string
      white_list:
        description: 白名单ip列表
        type: string
    required:
    - ip_list
    - port
    - service_name
    - weight_list
    type: object
  dto.CreateOrUpdateHttpServiceReq:
    properties:
      black_list:
        description: 黑名单ip列表
        type: string
      client_ip_flow_limit:
        description: 客户端ip限流
        type: integer
      header_transform:
        description: header转换
        type: string
      ip_list:
        description: ip列表
        type: string
      need_https:
        description: 支持https
        type: integer
      need_strip_uri:
        description: 启用strip_uri
        type: integer
      need_websocket:
        description: 支持websocket
        type: integer
      open_auth:
        description: 是否开启权限
        type: integer
      round_type:
        description: 轮询方式
        type: integer
      rule:
        description: 域名或者前缀
        type: string
      rule_type:
        description: 接入类型
        type: integer
      service_desc:
        description: 服务描述
        type: string
      service_host_flow_limit:
        description: 服务端主机限流
        type: integer
      service_name:
        description: 服务名称
        type: string
      upstream_connect_timeout:
        description: 建立连接超时时间, 单位为s
        type: integer
      upstream_header_timeout:
        description: 获取header超时时间, 单位为s
        type: integer
      upstream_idle_timeout:
        description: 连接最大空闲时间, 单位为s
        type: integer
      upstream_max_idle:
        description: 最大连接空闲数量
        type: integer
      url_rewrite:
        description: 启用url重写
        type: string
      weight_list:
        description: 权重列表
        type: string
      white_list:
        description: 白名单ip列表
        type: string
    required:
    - ip_list
    - rule
    - service_name
    - weight_list
    type: object
  dto.CreateOrUpdateTcpServiceReq:
    properties:
      black_list:
        description: 黑名单ip列表
        type: string
      client_ip_flow_limit:
        description: 客户端ip限流
        type: integer
      forbid_list:
        description: 禁用ip列表
        type: string
      ip_list:
        description: ip列表
        type: string
      open_auth:
        description: 是否开启权限
        type: integer
      port:
        description: 端口
        type: integer
      round_type:
        description: 轮询方式
        type: integer
      service_desc:
        description: 服务描述
        type: string
      service_host_flow_limit:
        description: 服务端主机限流
        type: integer
      service_name:
        description: 服务名称
        type: string
      weight_list:
        description: 权重列表
        type: string
      white_host_name:
        description: 白名单主机列表
        type: string
      white_list:
        description: 白名单ip列表
        type: string
    required:
    - ip_list
    - port
    - service_name
    - weight_list
    type: object
  dto.Flow:
    properties:
      today_flow:
        description: 今日流量列表
        items:
          type: integer
        type: array
      yesterday_flow:
        description: 昨日流量列表
        items:
          type: integer
        type: array
    type: object
...
  dto.ListServiceItem:
    properties:
      id:
        type: integer
      real_qpd:
        type: integer
      real_qps:
        type: integer
      service_addr:
        type: string
      service_desc:
        type: string
      service_name:
        type: string
      service_type:
        type: integer
      total_node:
        type: integer
    type: object
  dto.ListServicesRes:
    properties:
      items:
        description: 服务列表
        items:
          $ref: '#/definitions/dto.ListServiceItem'
        type: array
      total:
        description: 共计条数
        example: 0
        type: integer
    type: object
...
  dto.UserInfo:
    properties:
      avatar:
        type: string
      id:
        type: integer
      intro:
        type: string
      login_at:
        type: string
      roles:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
  po.AccessControl:
    properties:
      black_list:
        type: string
      client_ip_flow_limit:
        description: BlackHostName     string `json:"black_host_name" gorm:"column:black_host_name"
          description:"黑名单主机"`
        type: integer
      id:
        type: integer
      open_auth:
        type: integer
      service_host_flow_limit:
        type: integer
      service_id:
        type: integer
      white_host_name:
        type: string
      white_list:
        type: string
    type: object
...
  po.GrpcRule:
    properties:
      header_transform:
        type: string
      id:
        type: integer
      port:
        type: integer
      service_id:
        type: integer
    type: object
  po.HttpRule:
    properties:
      header_transform:
        type: string
      id:
        type: integer
      need_https:
        type: integer
      need_strip_uri:
        type: integer
      need_websocket:
        type: integer
      rule:
        type: string
      rule_type:
        type: integer
      service_id:
        type: integer
      url_rewrite:
        type: string
    type: object
  po.LoadBalance:
    properties:
      check_interval:
        type: integer
      check_method:
        type: integer
      check_timeout:
        type: integer
      forbid_list:
        type: string
      id:
        type: integer
      ip_list:
        type: string
      round_type:
        type: integer
      service_id:
        type: integer
      upstream_connect_timeout:
        type: integer
      upstream_header_timeout:
        type: integer
      upstream_idle_timeout:
        type: integer
      upstream_max_idle:
        type: integer
      weight_list:
        type: string
    type: object
  po.ServiceDetail:
    properties:
      access_control:
        $ref: '#/definitions/po.AccessControl'
      grpc_rule:
        $ref: '#/definitions/po.GrpcRule'
      http_rule:
        $ref: '#/definitions/po.HttpRule'
      info:
        $ref: '#/definitions/po.ServiceInfo'
      load_balance:
        $ref: '#/definitions/po.LoadBalance'
      tcp_rule:
        $ref: '#/definitions/po.TcpRule'
    type: object
  po.ServiceInfo:
    properties:
      create_at:
        type: string
      id:
        type: integer
      is_delete:
        type: integer
      service_desc:
        type: string
      service_name:
        type: string
      service_type:
        type: integer
      update_at:
        type: string
    type: object
  po.TcpRule:
    properties:
      id:
        type: integer
      port:
        type: integer
      service_id:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /dashboard/flow/services/{service_id}:
    get:
      description: 查询服务流量
      operationId: /flow/services/{service_id}
      parameters:
      - description: service id
        in: path
        name: service_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.Flow'
              type: object
      summary: 查询服务流量接口
      tags:
      - 数据接口
...
  /services:
    get:
      description: 查询服务列表
      operationId: /services
      parameters:
      - description: keyword
        in: query
        name: keyword
        type: string
      - description: page index
        in: query
        name: page_index
        type: string
      - description: page size
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ListServicesRes'
              type: object
      summary: 查询服务列表接口
      tags:
      - 服务接口
  /services/{service_id}:
    delete:
      description: 删除服务
      operationId: /services/{service_id}
      parameters:
      - description: service id
        in: path
        name: service_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 删除服务接口
      tags:
      - 服务接口
    get:
      description: 查询服务详情
      operationId: /services/{service_id}
      parameters:
      - description: service id
        in: path
        name: service_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/po.ServiceDetail'
              type: object
      summary: 查询服务详情接口
      tags:
      - 服务接口
  /services/grpc:
    post:
      consumes:
      - application/json
      description: 创建grpc服务
      operationId: /services/grpc
      parameters:
      - description: create grpc service request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrUpdateGrpcServiceReq'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/po.ServiceDetail'
              type: object
      summary: 创建grpc服务接口
      tags:
      - 服务接口
  /services/grpc/{service_id}:
    put:
      consumes:
      - application/json
      description: 更新grpc服务
      operationId: /services/grpc/{service_id}
      parameters:
      - description: service id
        in: path
        name: service_id
        required: true
        type: string
      - description: update grpc service request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrUpdateGrpcServiceReq'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/po.ServiceDetail'
              type: object
      summary: 更新grpc服务接口
      tags:
      - 服务接口
  /services/http:
    post:
      consumes:
      - application/json
      description: 创建http服务
      operationId: /services/http
      parameters:
      - description: create http service request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrUpdateHttpServiceReq'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/po.ServiceDetail'
              type: object
      summary: 创建http服务接口
      tags:
      - 服务接口
  /services/http/{service_id}:
    put:
      consumes:
      - application/json
      description: 更新http服务
      operationId: /services/http/{service_id}
      parameters:
      - description: service id
        in: path
        name: service_id
        required: true
        type: string
      - description: update http service request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrUpdateHttpServiceReq'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/po.ServiceDetail'
              type: object
      summary: 更新http服务接口
      tags:
      - 服务接口
  /services/tcp:
    post:
      consumes:
      - application/json
      description: 创建tcp服务
      operationId: /services/tcp
      parameters:
      - description: create tcp service request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrUpdateTcpServiceReq'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/po.ServiceDetail'
              type: object
      summary: 创建tcp服务接口
      tags:
      - 服务接口
  /services/tcp/{service_id}:
    put:
      consumes:
      - application/json
      description: 更新tcp服务
      operationId: /services/tcp/{service_id}
      parameters:
      - description: service id
        in: path
        name: service_id
        required: true
        type: string
      - description: update tcp service request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrUpdateTcpServiceReq'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/po.ServiceDetail'
              type: object
      summary: 更新tcp服务接口
      tags:
      - 服务接口
  /users/admin:
    get:
      description: 获取用户的信息
      operationId: /users/admin
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserInfo'
              type: object
      summary: 查询用户信息接口
      tags:
      - 用户接口
    patch:
      consumes:
      - application/json
      description: 修改用户的密码
      operationId: /users/admin
      parameters:
      - description: change user password request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangeUserPwdReq'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/common_middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 修改用户密码接口
      tags:
      - 用户接口
swagger: "2.0"